refactor(room): conver DAO to Repository:
