+++ room/src/infra/db/db_daoimpl_question.go
@@ -6,33 +6,33 @@ import (
 	"gopkg.in/mgo.v2/bson"
 )
 
-type QuestionDaoImpl struct {
+type QuestionRepositoryImpl struct {
 	collection *mgo.Collection
 }
 
-func NewQuestionDaoImpl() *QuestionDaoImpl {
-	return &QuestionDaoImpl{DB.C(models.Question{}.TableName())}
+func NewQuestionRepositoryImpl() *QuestionRepositoryImpl {
+	return &QuestionRepositoryImpl{DB.C(models.Question{}.TableName())}
 }
 
-func (daoi QuestionDaoImpl) CreateQuestion(question *models.Question) error {
+func (daoi QuestionRepositoryImpl) CreateQuestion(question *models.Question) error {
 	err := daoi.collection.Insert(question)
 	return err
 }
 
-func (daoi QuestionDaoImpl) GetQuestionById(id string) (models.Question, error) {
+func (daoi QuestionRepositoryImpl) GetQuestionById(id string) (models.Question, error) {
 	question := models.Question{}
 	err := daoi.collection.Find(bson.M{"id": id}).One(&question)
 	return question, err
 
 }
 
-func (daoi QuestionDaoImpl) GetQuestionsWithDifficulty(difficulty models.QuestionDifficulty) ([]models.Question, error) {
+func (daoi QuestionRepositoryImpl) GetQuestionsWithDifficulty(difficulty models.QuestionDifficulty) ([]models.Question, error) {
 	var questions []models.Question
 	err := daoi.collection.Find(bson.M{"difficulty": difficulty}).All(&questions)
 	return questions, err
 }
 
-func (daoi QuestionDaoImpl) ClearQuestions() error {
+func (daoi QuestionRepositoryImpl) ClearQuestions() error {
 	err := daoi.collection.DropCollection()
 	return err
 }
