+++ room/src/repositories/repository_question.go
@@ -1,8 +1,8 @@
-package dao
+package repositories
 
 import "code2gather.com/room/src/models"
 
-type QuestionDAO interface {
+type QuestionRepository interface {
 	CreateQuestion(question *models.Question) error
 	GetQuestionById(id string) (models.Question, error)
 	GetQuestionsWithDifficulty(difficulty models.QuestionDifficulty) ([]models.Question, error)
