+++ room/src/repositories/repository_room.go
@@ -1,4 +1,4 @@
-package dao
+package repositories
 
 import (
 	"gopkg.in/mgo.v2/bson"
@@ -6,7 +6,7 @@ import (
 	"code2gather.com/room/src/models"
 )
 
-type RoomDAO interface {
+type RoomRepository interface {
 	CreateRoom(room *models.Room) error
 	GetRoomById(id string) (models.Room, error)
 	UpdateRoom(room *models.Room) error
